requirements.txt
scrapy>=2.11.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
anthropic>=0.18.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0